requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
azure-monitor-opentelemetry==1.2.0
typing-extensions>=4.8.0
pydantic>=2.4.0
setuptools>=68.0.0
requests>=2.31.0